minor structure changes and a _delete() method
worked on a Tablemodify._delete(instance) method to completely reset an
instance. It already works quite well except several event listeners are
not removed yet and the "native" order of rows gets lost... Therefore I
changed the structure of tm a bit:
now a module must return an object with at least one method: unset() to
revert changes of this module.
Additionally, all active modules  (these returned objects) of a
tm-instance are stored in a new property "activeModules"  now.