Major structural changes
removed the render() method. sorting etc now works "in place", so this
is no longer required.
added a couple of methods to the tm class, e.g. addRows() to dynamically
add rows, with correct sorting and filtering.
added signal() to trigger events on the original table, for
communication between modules.
renamed "rows" to "visibleRows" and "fragment" to "hiddenRows".
added some comments